Refactor the package structure.
